+++ src/stellium/presentation/__init__.py
@@ -21,7 +21,7 @@ Usage:
 from stellium.core.protocols import ReportRenderer, ReportSection
 
 from .builder import ReportBuilder
-from .renderers import HTMLRenderer, PlainTextRenderer, RichTableRenderer
+from .renderers import HTMLRenderer, PlainTextRenderer, ProseRenderer, RichTableRenderer
 
 # Import from the new sections package
 from .sections import (
@@ -46,4 +46,5 @@ __all__ = [
     "RichTableRenderer",
     "PlainTextRenderer",
     "HTMLRenderer",
+    "ProseRenderer",
 ]